requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
attrs==23.2.0
cdifflib==1.2.6
dotmap==1.3.30
frozenlist==1.4.1
idna==3.7
imageio==2.34.2
lazy_loader==0.4
multidict==6.0.5
networkx==3.3
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
scikit-image==0.24.0
scipy==1.14.0
tesserocr @ https://github.com/simonflueckiger/tesserocr-windows_build/releases/download/tesserocr-v2.7.0-tesseract-5.3.4/tesserocr-2.7.0-cp312-cp312-win_amd64.whl#sha256=93341971973e1bcf3ab1aa57f083b7298d77af1e3426a78bfd41a35abf22523c
tifffile==2024.7.24
yarl==1.9.4
